Sharing markdown parsing login between both middlewares #119
